booster: pin exec targets and add manpage
We use a number of libraries and a configuration path that are linked
during build, which also have binaries that get called at runtime.
Allowing these to not point to the same backing derivations is a great
way to introduce hard to reproduce failures. As such, we have patched
the source to exec fully qualified paths, as was previously done for
/usr/bin/fsck.

While were were here, there is also a nice manpage that we failed to
include before, so that has been resolved with help from roff. As well,
there are a number of /usr/something defaults that have been pointed to
their correct /nix/store locations.